pyproject.toml
[tool.poetry]
name = "personal website"
version = "0.1.0"
description = "Personal website in Django"
authors = ["Friday G <faradayyg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1.3"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
django-storages = "^1.14.4"
django-taggit = "^6.1.0"
boto3 = "^1.35.72"
python-decouple = "^3.8"

[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.3"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
